requirements.txt
blurhash==1.1.4
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
idna==3.4
Mastodon.py==1.5.2
mwparserfromhell==0.6.4
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.6
requests==2.28.1
six==1.16.0
urllib3==1.26.12
